// src/app/claim-processing/manual-claim-processing/process-claim-drugchrg/model/screenbean.model.ts
import {ScreenDrugChargeLines} from './screen-drug-charge-lines.model';
/**
 * Model class Screenbean
 * Path: screenbean/procclmdrugchrg
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmdrugchrg::Screenbean
 * Legacy Mapping: SCREEN-SAVE-AREA
 */
export class Screenbean {
  filler1 = '';
  mapCommandLine = '';
  filler2 = '';
  screenDrugChargeLines: ScreenDrugChargeLines[] = [];
  screenExPay1 = '';
  screenExPay2 = '';
  screenAfterTerm = '';
  screenErr1 = '';
  screenErr2 = '';
  screenErr3 = '';
  screenPf8 = '';
  m27cmnd = '';
  m27mem1 = '';
  m27mem2 = '';
  m27mem3 = '';
  m27cno1 = '';
  m27cno2 = '';
  m27cno3 = '';
  m27cno4 = '';
  m27cno5 = '';
  m27cno6 = '';
  m27com1 = '';
  m27com2 = '';
  m27name = '';
  m27memn = '';
  m27assc = '';
  m27pln1 = '';
  m27eff1 = '';
  m27trm1 = '';
  m27dob = '';
  m27pln2 = '';
  m27eff2 = '';
  m27trm2 = '';
  m27pln3 = '';
  m27eff3 = '';
  m27trm3 = '';
  m27pro1 = '';
  m27phy1 = '';
  m27sr11 = '';
  m27sr21 = '';
  m27chg1 = '';
  m27ndc1 = '';
  m27np1 = '';
  m27pl1 = '';
  m27ind1 = '';
  m27pre1 = '';
  m27pro2 = '';
  m27phy2 = '';
  m27sr12 = '';
  m27sr22 = '';
  m27chg2 = '';
  m27ndc2 = '';
  m27np2 = '';
  m27pl2 = '';
  m27ind2 = '';
  m27pre2 = '';
  m27pro3 = '';
  m27phy3 = '';
  m27sr13 = '';
  m27sr23 = '';
  m27chg3 = '';
  m27ndc3 = '';
  m27np3 = '';
  m27pl3 = '';
  m27ind3 = '';
  m27pre3 = '';
  m27pro4 = '';
  m27phy4 = '';
  m27sr14 = '';
  m27sr24 = '';
  m27chg4 = '';
  m27ndc4 = '';
  m27np4 = '';
  m27pl4 = '';
  m27ind4 = '';
  m27pre4 = '';
  m27pro5 = '';
  m27phy5 = '';
  m27sr15 = '';
  m27sr25 = '';
  m27chg5 = '';
  m27ndc5 = '';
  m27np5 = '';
  m27pl5 = '';
  m27ind5 = '';
  m27pre5 = '';
  m27pro6 = '';
  m27phy6 = '';
  m27sr16 = '';
  m27sr26 = '';
  m27chg6 = '';
  m27ndc6 = '';
  m27np6 = '';
  m27pl6 = '';
  m27ind6 = '';
  m27pre6 = '';
  m27pro7 = '';
  m27phy7 = '';
  m27sr17 = '';
  m27sr27 = '';
  m27chg7 = '';
  m27ndc7 = '';
  m27np7 = '';
  m27pl7 = '';
  m27ind7 = '';
  m27pre7 = '';
  m27pro8 = '';
  m27phy8 = '';
  m27sr18 = '';
  m27sr28 = '';
  m27chg8 = '';
  m27ndc8 = '';
  m27np8 = '';
  m27pl8 = '';
  m27ind8 = '';
  m27pre8 = '';
  m27xpa1 = '';
  m27xpa2 = '';
  m27atrm = '';
  m27err1 = '';
  m27err2 = '';
  m27err3 = '';
  m27pf8 = '';
}